Se aplicaron nuevos cambios en Detail Pinata

+++ MyTables/Piñatas_detalles.py
@@ -7,7 +7,7 @@ class piñatas_detalles(Model):
     Colores = CharField(max_length=100)
     Tiempo_estimado = CharField(max_length=45)
     Stock = IntegerField(primary_key=False)
-    Piñata_idPiñatas = ForeignKeyField(piñata, field= "idPiñatas", backref='piñatas_detalles', column_name='Piñata_idPiñatas')
+    Piñata_idPiñatas = IntegerField(ForeignKeyField(piñata, field= "idPiñatas", backref='piñatas_detalles', column_name='Piñata_idPiñatas'))
 
     def __int__(self):
         return self.idPiñatas_detalles
